Refactor item folder
